requirements.txt
anyio==4.3.0
certifi==2024.2.2
feedparser==6.0.11
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
python-telegram-bot==20.8
schedule==1.2.1
sgmllib3k==1.0.0
sniffio==1.3.1
